Fixed a few bugs

+++ Assets/Scriptable Objects/ScriptableGuns.cs
@@ -7,7 +7,7 @@ public class ScriptableGuns : ScriptableObject
     [Header("Information")]
     public string gunName;
     public string gunCategory;
-    public float gunPrice;
+    public int gunPrice;
     public float roundCapacity;
 
     [Header("Model")]
